Revert unnecessary changes

+++ source/com.kms.katalon.core.webui/src/com/kms/katalon/core/webui/common/ScreenUtil.java
@@ -28,7 +28,7 @@ public class ScreenUtil {
         mainScreen = new DesktopScreenRegion();
     }
 
-    public ScreenUtil(int similarity) {
+    public ScreenUtil(double similarity) {
         this();
         this.similarity = similarity;
     }
